+++ src/constants/IOSCategory.ts
@@ -1,0 +1,8 @@
+export enum IOSCategory {
+  Playback = 'playback',
+  PlayAndRecord = 'playAndRecord',
+  MultiRoute = 'multiRoute',
+  Ambient = 'ambient',
+  SoloAmbient = 'soloAmbient',
+  Record = 'record',
+}
