+++ src/constants/IOSCategoryMode.ts
@@ -1,0 +1,11 @@
+export enum IOSCategoryMode {
+  Default = 'default',
+  GameChat = 'gameChat',
+  Measurement = 'measurement',
+  MoviePlayback = 'moviePlayback',
+  SpokenAudio = 'spokenAudio',
+  VideoChat = 'videoChat',
+  VideoRecording = 'videoRecording',
+  VoiceChat = 'voiceChat',
+  VoicePrompt = 'voicePrompt',
+}
