+++ src/constants/IOSCategoryOptions.ts
@@ -1,0 +1,9 @@
+export enum IOSCategoryOptions {
+  MixWithOthers = 'mixWithOthers',
+  DuckOthers = 'duckOthers',
+  InterruptSpokenAudioAndMixWithOthers = 'interruptSpokenAudioAndMixWithOthers',
+  AllowBluetooth = 'allowBluetooth',
+  AllowBluetoothA2DP = 'allowBluetoothA2DP',
+  AllowAirPlay = 'allowAirPlay',
+  DefaultToSpeaker = 'defaultToSpeaker',
+}
