+++ src/index.ts
@@ -1,5 +1,6 @@
-import TrackPlayer from './trackPlayer'
+import * as TrackPlayer from './trackPlayer'
 
+export * from './constants';
 export * from './hooks'
 export * from './interfaces'
