+++ src/interfaces/Progress.ts
@@ -1,0 +1,5 @@
+export interface Progress {
+  position: number
+  duration: number
+  buffered: number
+}
