+++ src/interfaces/ResourceObject.ts
@@ -1,0 +1,1 @@
+export type ResourceObject = number
