+++ src/interfaces/ServiceHandler.ts
@@ -1,0 +1,1 @@
+export type ServiceHandler = () => Promise<void>
