+++ src/interfaces/events/PlaybackErrorEvent.ts
@@ -1,0 +1,4 @@
+export interface PlaybackErrorEvent {
+  code: string
+  message: string
+}
