+++ src/interfaces/events/PlaybackProgressUpdatedEvent.ts
@@ -1,0 +1,5 @@
+import type { Progress } from '../Progress'
+
+export interface PlaybackProgressUpdatedEvent extends Progress {
+  track: number
+}
