+++ src/interfaces/events/PlaybackQueueEndedEvent.ts
@@ -1,0 +1,4 @@
+export interface PlaybackQueueEndedEvent {
+  track: number
+  position: number
+}
