+++ src/interfaces/events/PlaybackStateEvent.ts
@@ -1,0 +1,5 @@
+import type { State } from '../../constants'
+
+export interface PlaybackStateEvent {
+  state: State
+}
