+++ src/interfaces/events/PlaybackTrackChangedEvent.ts
@@ -1,0 +1,5 @@
+export interface PlaybackTrackChangedEvent {
+  track: number
+  position: number
+  nextTrack: number
+}
