+++ src/interfaces/events/RemoteDuckEvent.ts
@@ -1,0 +1,4 @@
+export interface RemoteDuckEvent {
+  paused: boolean
+  permanent: boolean
+}
