+++ src/interfaces/events/RemoteJumpBackwardEvent.ts
@@ -1,0 +1,3 @@
+export interface RemoteJumpBackwardEvent {
+  interval: number
+}
