+++ src/interfaces/events/RemoteJumpForwardEvent.ts
@@ -1,0 +1,3 @@
+export interface RemoteJumpForwardEvent {
+  interval: number
+}
