+++ src/interfaces/events/RemotePlayIdEvent.ts
@@ -1,0 +1,3 @@
+export interface RemotePlayIdEvent {
+  id: string
+}
