+++ src/interfaces/events/RemotePlaySearchEvent.ts
@@ -1,0 +1,9 @@
+export interface RemotePlaySearchEvent {
+  query: string
+  focus: string
+  title: string
+  artist: string
+  album: string
+  date: string
+  playlist: string
+}
