+++ src/interfaces/events/RemoteSeekEvent.ts
@@ -1,0 +1,3 @@
+export interface RemoteSeekEvent {
+  position: number
+}
