+++ src/interfaces/events/RemoteSetRatingEvent.ts
@@ -1,0 +1,3 @@
+export interface RemoteSetRatingEvent {
+  rating: boolean | number
+}
