+++ src/interfaces/events/RemoteSkipEvent.ts
@@ -1,0 +1,3 @@
+export interface RemoteSkipEvent {
+  index: number
+}
