+++ src/interfaces/index.ts
@@ -1,0 +1,10 @@
+export * from './FeedbackOptions'
+export * from './MetadataOptions'
+export * from './NowPlayingMetadata'
+export * from './PlayerOptions'
+export * from './Progress'
+export * from './ResourceObject'
+export * from './ServiceHandler'
+export * from './Track'
+export * from './TrackMetadataBase'
+export * from './events'
