+++ src/trackPlayer.ts
@@ -1,14 +1,13 @@
 import { AppRegistry, DeviceEventEmitter, NativeEventEmitter, NativeModules, Platform } from 'react-native'
 // @ts-ignore
-import * as resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource'
-import {
-  Event,
+import resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource'
+import type { Event, RepeatMode, State } from './constants'
+import type {
   EventPayloadByEvent,
   MetadataOptions,
   NowPlayingMetadata,
   PlayerOptions,
-  RepeatMode,
-  State,
+  ServiceHandler,
   Track,
   TrackMetadataBase,
 } from './interfaces'
@@ -17,7 +16,6 @@ const { TrackPlayerModule: TrackPlayer } = NativeModules
 const emitter = Platform.OS !== 'android' ? new NativeEventEmitter(TrackPlayer) : DeviceEventEmitter
 
 // MARK: - Helpers
-
 function resolveImportedPath(path?: number | string) {
   if (!path) return undefined
   return resolveAssetSource(path) || path
@@ -34,15 +32,14 @@ function optionalNumberToDefault(num?: number, defaultValue: number = -1): numbe
 /**
  * Initializes the player with the specified options.
  */
-async function setupPlayer(options: PlayerOptions = {}): Promise<void> {
+export async function setupPlayer(options: PlayerOptions = {}): Promise<void> {
   return TrackPlayer.setupPlayer(options || {})
 }
 
-type ServiceHandler = () => Promise<void>
 /**
  * Register the playback service. The service will run as long as the player runs.
  */
-function registerPlaybackService(factory: () => ServiceHandler) {
+export function registerPlaybackService(factory: () => ServiceHandler) {
   if (Platform.OS === 'android') {
     // Registers the headless task
     AppRegistry.registerHeadlessTask('TrackPlayer', factory)
@@ -52,18 +49,11 @@ function registerPlaybackService(factory: () => ServiceHandler) {
   }
 }
 
-function addEventListener<T extends Event>(
+export function addEventListener<T extends Event>(
   event: T,
   listener: EventPayloadByEvent[T] extends never ? () => void : (event: EventPayloadByEvent[T]) => void,
 ) {
   return emitter.addListener(event, listener)
-}
-
-/**
- * @deprecated This method should not be used, most methods reject when service is not bound.
- */
-function isServiceRunning(): Promise<boolean> {
-  return TrackPlayer.isServiceRunning()
 }
 
 // MARK: - Queue API
@@ -71,7 +61,7 @@ function isServiceRunning(): Promise<boolean> {
 /**
  * Adds one or more tracks to the queue.
  */
-async function add(tracks: Track | Track[], insertBeforeIndex?: number): Promise<number | void> {
+export async function add(tracks: Track | Track[], insertBeforeIndex?: number): Promise<number | void> {
   // Clone the array before modifying it
   if (Array.isArray(tracks)) {
     tracks = [...tracks]
@@ -97,7 +87,7 @@ async function add(tracks: Track | Track[], insertBeforeIndex?: number): Promise
 /**
  * Removes one or more tracks from the queue.
  */
-async function remove(tracks: number | number[]): Promise<void> {
+export async function remove(tracks: number | number[]): Promise<void> {
   if (!Array.isArray(tracks)) {
     tracks = [tracks]
   }
@@ -108,28 +98,28 @@ async function remove(tracks: number | number[]): Promise<void> {
 /**
  * Clears any upcoming tracks from the queue.
  */
-async function removeUpcomingTracks(): Promise<void> {
+export async function removeUpcomingTracks(): Promise<void> {
   return TrackPlayer.removeUpcomingTracks()
 }
 
 /**
  * Skips to a track in the queue.
  */
-async function skip(trackIndex: number, initialPosition?: number): Promise<void> {
+export async function skip(trackIndex: number, initialPosition?: number): Promise<void> {
   return TrackPlayer.skip(trackIndex, optionalNumberToDefault(initialPosition))
 }
 
 /**
  * Skips to the next track in the queue.
  */
-async function skipToNext(initialPosition?: number): Promise<void> {
+export async function skipToNext(initialPosition?: number): Promise<void> {
   return TrackPlayer.skipToNext(optionalNumberToDefault(initialPosition))
 }
 
 /**
  * Skips to the previous track in the queue.
  */
-async function skipToPrevious(initialPosition?: number): Promise<void> {
+export async function skipToPrevious(initialPosition?: number): Promise<void> {
   return TrackPlayer.skipToPrevious(optionalNumberToDefault(initialPosition))
 }
 
@@ -138,7 +128,7 @@ async function skipToPrevious(initialPosition?: number): Promise<void> {
 /**
  * Updates the configuration for the components.
  */
-async function updateOptions(options: MetadataOptions = {}): Promise<void> {
+export async function updateOptions(options: MetadataOptions = {}): Promise<void> {
   options = { ...options }
 
   // Resolve the asset for each icon
@@ -158,7 +148,7 @@ async function updateOptions(options: MetadataOptions = {}): Promise<void> {
  * Updates the metadata of a track in the queue. If the current track is updated,
  * the notification and the Now Playing Center will be updated accordingly.
  */
-async function updateMetadataForTrack(trackIndex: number, metadata: TrackMetadataBase): Promise<void> {
+export async function updateMetadataForTrack(trackIndex: number, metadata: TrackMetadataBase): Promise<void> {
   // Clone the object before modifying it
   metadata = Object.assign({}, metadata)
 
@@ -168,11 +158,11 @@ async function updateMetadataForTrack(trackIndex: number, metadata: TrackMetadat
   return TrackPlayer.updateMetadataForTrack(trackIndex, metadata)
 }
 
-function clearNowPlayingMetadata(): Promise<void> {
+export function clearNowPlayingMetadata(): Promise<void> {
   return TrackPlayer.clearNowPlayingMetadata()
 }
 
-function updateNowPlayingMetadata(metadata: NowPlayingMetadata): Promise<void> {
+export function updateNowPlayingMetadata(metadata: NowPlayingMetadata): Promise<void> {
   // Clone the object before modifying it
   metadata = Object.assign({}, metadata)
 
@@ -187,49 +177,49 @@ function updateNowPlayingMetadata(metadata: NowPlayingMetadata): Promise<void> {
 /**
  * Resets the player stopping the current track and clearing the queue.
  */
-async function reset(): Promise<void> {
+export async function reset(): Promise<void> {
   return TrackPlayer.reset()
 }
 
 /**
  * Plays or resumes the current track.
  */
-async function play(): Promise<void> {
+export async function play(): Promise<void> {
   return TrackPlayer.play()
 }
 
 /**
  * Pauses the current track.
  */
-async function pause(): Promise<void> {
+export async function pause(): Promise<void> {
   return TrackPlayer.pause()
 }
 
 /**
  * Seeks to a specified time position in the current track.
  */
-async function seekTo(position: number): Promise<void> {
+export async function seekTo(position: number): Promise<void> {
   return TrackPlayer.seekTo(position)
 }
 
 /**
  * Sets the volume of the player.
  */
-async function setVolume(level: number): Promise<void> {
+export async function setVolume(level: number): Promise<void> {
   return TrackPlayer.setVolume(level)
 }
 
 /**
  * Sets the playback rate.
  */
-async function setRate(rate: number): Promise<void> {
+export async function setRate(rate: number): Promise<void> {
   return TrackPlayer.setRate(rate)
 }
 
 /**
  * Sets the repeat mode.
  */
-async function setRepeatMode(mode: RepeatMode): Promise<RepeatMode> {
+export async function setRepeatMode(mode: RepeatMode): Promise<RepeatMode> {
   return TrackPlayer.setRepeatMode(mode)
 }
 
@@ -238,112 +228,69 @@ async function setRepeatMode(mode: RepeatMode): Promise<RepeatMode> {
 /**
  * Gets the volume of the player (a number between 0 and 1).
  */
-async function getVolume(): Promise<number> {
+export async function getVolume(): Promise<number> {
   return TrackPlayer.getVolume()
 }
 
 /**
  * Gets the playback rate, where 1 is the regular speed.
  */
-async function getRate(): Promise<number> {
+export async function getRate(): Promise<number> {
   return TrackPlayer.getRate()
 }
 
 /**
  * Gets a track object from the queue.
  */
-async function getTrack(trackIndex: number): Promise<Track | null> {
+export async function getTrack(trackIndex: number): Promise<Track | null> {
   return TrackPlayer.getTrack(trackIndex)
 }
 
 /**
  * Gets the whole queue.
  */
-async function getQueue(): Promise<Track[]> {
+export async function getQueue(): Promise<Track[]> {
   return TrackPlayer.getQueue()
 }
 
 /**
  * Gets the index of the current track.
  */
-async function getCurrentTrack(): Promise<number> {
+export async function getCurrentTrack(): Promise<number> {
   return TrackPlayer.getCurrentTrack()
 }
 
 /**
  * Gets the duration of the current track in seconds.
  */
-async function getDuration(): Promise<number> {
+export async function getDuration(): Promise<number> {
   return TrackPlayer.getDuration()
 }
 
 /**
  * Gets the buffered position of the player in seconds.
  */
-async function getBufferedPosition(): Promise<number> {
+export async function getBufferedPosition(): Promise<number> {
   return TrackPlayer.getBufferedPosition()
 }
 
 /**
  * Gets the position of the player in seconds.
  */
-async function getPosition(): Promise<number> {
+export async function getPosition(): Promise<number> {
   return TrackPlayer.getPosition()
 }
 
 /**
  * Gets the state of the player.
  */
-async function getState(): Promise<State> {
+export async function getState(): Promise<State> {
   return TrackPlayer.getState()
 }
 
 /**
  * Gets the repeat mode.
  */
-async function getRepeatMode(): Promise<RepeatMode> {
+export async function getRepeatMode(): Promise<RepeatMode> {
   return TrackPlayer.getRepeatMode()
-}
-
-export default {
-  // MARK: - General API
-  setupPlayer,
-  registerPlaybackService,
-  addEventListener,
-  isServiceRunning,
-
-  // MARK: - Queue API
-  add,
-  remove,
-  removeUpcomingTracks,
-  skip,
-  skipToNext,
-  skipToPrevious,
-
-  // MARK: - Control Center / Notifications API
-  updateOptions,
-  updateMetadataForTrack,
-  clearNowPlayingMetadata,
-  updateNowPlayingMetadata,
-
-  // MARK: - Player API
-  reset,
-  play,
-  pause,
-  seekTo,
-  setVolume,
-  setRate,
-  setRepeatMode,
-
-  // MARK: - Getters
-  getVolume,
-  getRate,
-  getTrack,
-  getQueue,
-  getCurrentTrack,
-  getDuration,
-  getBufferedPosition,
-  getPosition,
-  getState,
-  getRepeatMode,
 }
